+++ src/EngineCore/FMT/Context/FormatterContext/FormatterContextArgsTuple.h
@@ -124,7 +124,7 @@ namespace EngineCore::FMT::Detail {
             {}
             
         public:
-            int Size() override
+            size_t Size() override
                     { return m_contextArgs.Size(); }
 
             void RunTypeAtIndex(const Detail::FormatIndex& idx) override
